[refactor]: Refactor layout
